+++ release/components/datatable.component.js
@@ -512,6 +512,7 @@ var DatatableComponent = (function () {
         get: function () {
             return this.selected &&
                 this.rows &&
+                this.rows.length !== 0 &&
                 this.selected.length === this.rows.length;
         },
         enumerable: true,